Delete the getURLsSbeOrRelative(url) function
This function was only used for tools.xml, and didnt even do its job...
As of this writing, it simply prepended the `baseURL` variable, which is
determined by the Snap! server, not s.b.e.

If we want to have s.b.e as a permanent fallback source we should re-implement
this in a way that is used for all Snap! media.