Update LICENSE file so it gets properly displayed
GitHub treats LICENSE files in a particular fashion. Whenever this file
is present in a repository, an icon is displayed at the top of the
repository page. This commit enables this behavior.

LICENSE text itself was updated to mirror the one from gnu.org website,
which had some differences such as https everywhere.


Former-commit-id: 64ac63776fc3e03e04f5e8a7815fb8a93f458deb