Refactor contributing guidelines
Having contributing guidelines in a single file can improve readability
for new users. GitHub also has a mechanism that invites users to read (or
re-read) contributing guidelines before submitting an issue or a modification.

This is why I merged the code and the translating guidelines into a
single file.

See GitHub blog article for reference:
https://help.github.com/articles/setting-guidelines-for-repository-contributors/

Former-commit-id: a0fe3bc249676fb495ad43f05959b67a4273d7eb