new experimental "When I am stopped" event option
scripts attached to this hat block will run when the user physically clicks on the red stop button or hits the esc-key on the keyboard for exactly one single atomic step ("frame") and then terminate. This means that e.g. only the first pass of a loop gets run. You can, however, use a WARP around the loop to force more iterations. any broadcasts or forked threads will likewise run for exactly a single step within the main one (i.e. the whole operation lasts exactly one global atom) This is experimental and might not make it into the release after all.


Former-commit-id: d11ba70ac6840781f064cbf1e2866e2740207a15