Provide IME support to the text editing in Snap.
1. Design Outline:

Three parties are involved:
- the target: the StringMorph/TextMorph to be edited
- the editor: the hidden textarea
- the manager: the CursorMorph

The manager takes care of the lifecycle of the editor, and provide visual
presentation of the cursor.

The editor holds the content being edited, handles keyboard events, perform
the required editing, then update the target with the result of editing.

The target handles mouse events, which result into changes of cursor position
and selection status, then it will update the editor with these information.

2. Change of WorldMorph.edit(target) method

At the start of the method, added two guards to check:
- if `target` is actually currently being edited, do nothing.
- if another morph is being edited, call stopEditing().
- ... the old codes

This is to fix a bug in mouse event handling. The bug is that shift+click does not
select the text between the clicked position and previous position, instead
all text wiil be selected.

The cause of this bug is the morph that being edited called stopEditing() and then
.edit(), which starts a new editing session, and the previous selection state is lost.
By adding above mentioned guard, the force restart will be avoid.


Former-commit-id: aabdfeade6ffb221dadf4146b0244e20a76be41f