Translate to Portuguese (Portugal) the audioComp_module library
Atention! There seem to be problems with the reporter for getting the stage size (width and hight). Opening in portuguese, the second menu contained "Stage", instead of "o palco", which broke the sound plotting command. I changed it back to "o palco", but it seems this was what got writen to the XML file. I edited this file and changed again all references to "o palco" to "Stage".


Former-commit-id: 47ebcd48e91dc975c1363caf2e19b61382f34a6b