Accounts for the scrollbars in ScrollFrameMorphs.
This ensures that the stopping point of the scrollbar is beyond the length of
the contents that it is scrolling, and that when scrolling is calculated, it
will now account for the presence of a scrollbar. This allows
you to scroll a little extra so you can always see all content.

Fixes #2468
Fixes #638
Fixes #804


Former-commit-id: b7bf8bab1a729354deae4083fec0a96d6eba7ecf