compiled multimap
This PR makes the compiled version of MAP variadic.  The expected use case is that the function input's ring does not have formal parameters, and  corresponding items of the input lists fill the empty slots in the ringed expression.

For compatibility, if the ring is given formal parameters, then the first one ("value") is bound to a list of the corresponding items of the input lists.


Former-commit-id: 35c7560816287dc0f7a2309f88ec532bc92b4b14