Enabled panning via with native 'balance' parameter
I edited the Tone block in the TuneScope library to pass the native
'balance' parameter, which feeds into a StereoPanningNode in the
TuneScope backend code. Additionally, we edited the TuneScope blocks
to remove 'Initialize TuneScope' from their block defintions; the user
will have to click this block at the start of the program once to load
the necessary dependencies.

We made these changes to better integrate TuneScope into the native
Snap! code. In the future, we would like to run the 'Initialize
Tunescope' block at load time, so that the user isn't tasked with
running it manually everytime they want to use the library.


Former-commit-id: 7e0394c098352981b93b7156185aa820e1a5d4a0