Remove all cache-busting attempts from IDE getURL
The Cache-Control header is disallowed by CORS on some servers, eg GitHub Pages
which then causes the entire request to fail.

It seems that at this point, we should not include any cache-busting code in the "raw" getURL function.
Years ago, there were unspecified "issues" with Finch/Hummingbird servers, the recent attempt
tried to avoid that by restricting to HTTPS but this is causing different troubles.

Therefore, attempts to cache bust should have clients append query string timestamps to URLs which are unlikely to cause problems with servers and can be done on a per-request basis.


Former-commit-id: f6bb1369d408ed17c8e99d5d68cb9620920ac454